Move post/curl/fetchUrl/fetchUrlFull to own class "Network\HTTPRequest"
