+++ mod/dfrn_request.php
@@ -39,6 +39,7 @@ use Friendica\Model\Notify\Type;
 use Friendica\Model\Profile;
 use Friendica\Model\User;
 use Friendica\Module\Security\Login;
+use Friendica\Network\HTTPRequest;
 use Friendica\Network\Probe;
 use Friendica\Protocol\Activity;
 use Friendica\Util\DateTimeFormat;
@@ -203,7 +204,7 @@ function dfrn_request_post(App $a)
 				}
 
 				if (!empty($dfrn_request) && strlen($confirm_key)) {
-					Network::fetchUrl($dfrn_request . '?confirm_key=' . $confirm_key);
+					HTTPRequest::fetchUrl($dfrn_request . '?confirm_key=' . $confirm_key);
 				}
 
 				// (ignore reply, nothing we can do it failed)
