+++ mod/match.php
@@ -27,6 +27,7 @@ use Friendica\Database\DBA;
 use Friendica\DI;
 use Friendica\Model\Contact;
 use Friendica\Model\Profile;
+use Friendica\Network\HTTPRequest;
 use Friendica\Util\Network;
 use Friendica\Util\Proxy as ProxyUtils;
 
@@ -76,7 +77,7 @@ function match_content(App $a)
 		$host = DI::baseUrl();
 	}
 
-	$msearch_json = Network::post($host . '/msearch', $params)->getBody();
+	$msearch_json = HTTPRequest::post($host . '/msearch', $params)->getBody();
 
 	$msearch = json_decode($msearch_json);
