+++ mod/parse_url.php
@@ -27,6 +27,7 @@ use Friendica\App;
 use Friendica\Core\Hook;
 use Friendica\Core\Logger;
 use Friendica\Core\System;
+use Friendica\Network\HTTPRequest;
 use Friendica\Util\Network;
 use Friendica\Util\ParseUrl;
 use Friendica\Util\Strings;
@@ -84,7 +85,7 @@ function parse_url_content(App $a)
 	// Check if the URL is an image, video or audio file. If so format
 	// the URL with the corresponding BBCode media tag
 	// Fetch the header of the URL
-	$curlResponse = Network::curl($url, false, ['novalidate' => true, 'nobody' => true]);
+	$curlResponse = HTTPRequest::curl($url, false, ['novalidate' => true, 'nobody' => true]);
 
 	if ($curlResponse->isSuccess()) {
 		// Convert the header fields into an array
