+++ mod/pubsubhubbub.php
@@ -24,6 +24,7 @@ use Friendica\Core\Logger;
 use Friendica\Database\DBA;
 use Friendica\DI;
 use Friendica\Model\PushSubscriber;
+use Friendica\Network\HTTPRequest;
 use Friendica\Util\Network;
 use Friendica\Util\Strings;
 
@@ -126,7 +127,7 @@ function pubsubhubbub_init(App $a) {
 		$hub_callback = rtrim($hub_callback, ' ?&#');
 		$separator = parse_url($hub_callback, PHP_URL_QUERY) === null ? '?' : '&';
 
-		$fetchResult = Network::fetchUrlFull($hub_callback . $separator . $params);
+		$fetchResult = HTTPRequest::fetchUrlFull($hub_callback . $separator . $params);
 		$body = $fetchResult->getBody();
 		$ret = $fetchResult->getReturnCode();
