+++ src/Core/Installer.php
@@ -27,6 +27,7 @@ use Friendica\Core\Config\Cache;
 use Friendica\Database\Database;
 use Friendica\Database\DBStructure;
 use Friendica\DI;
+use Friendica\Network\HTTPRequest;
 use Friendica\Util\Images;
 use Friendica\Util\Network;
 use Friendica\Util\Strings;
@@ -548,11 +549,11 @@ class Installer
 		$help = "";
 		$error_msg = "";
 		if (function_exists('curl_init')) {
-			$fetchResult = Network::fetchUrlFull($baseurl . "/install/testrewrite");
+			$fetchResult = HTTPRequest::fetchUrlFull($baseurl . "/install/testrewrite");
 
 			$url = Strings::normaliseLink($baseurl . "/install/testrewrite");
 			if ($fetchResult->getReturnCode() != 204) {
-				$fetchResult = Network::fetchUrlFull($url);
+				$fetchResult = HTTPRequest::fetchUrlFull($url);
 			}
 
 			if ($fetchResult->getReturnCode() != 204) {
