+++ src/Core/Worker.php
@@ -25,6 +25,7 @@ use Friendica\Core;
 use Friendica\Database\DBA;
 use Friendica\DI;
 use Friendica\Model\Process;
+use Friendica\Network\HTTPRequest;
 use Friendica\Util\DateTimeFormat;
 use Friendica\Util\Network;
 
@@ -992,7 +993,7 @@ class Worker
 		}
 
 		$url = DI::baseUrl() . '/worker';
-		Network::fetchUrl($url, false, 1);
+		HTTPRequest::fetchUrl($url, false, 1);
 	}
 
 	/**
