+++ src/Model/User.php
@@ -33,6 +33,7 @@ use Friendica\Core\Worker;
 use Friendica\Database\DBA;
 use Friendica\DI;
 use Friendica\Model\TwoFactor\AppSpecificPassword;
+use Friendica\Network\HTTPRequest;
 use Friendica\Network\HTTPException\InternalServerErrorException;
 use Friendica\Object\Image;
 use Friendica\Util\Crypto;
@@ -839,7 +840,7 @@ class User
 			$photo_failure = false;
 
 			$filename = basename($photo);
-			$img_str = Network::fetchUrl($photo, true);
+			$img_str = HTTPRequest::fetchUrl($photo, true);
 			// guess mimetype from headers or filename
 			$type = Images::guessType($photo, true);
