+++ src/Protocol/DFRN.php
@@ -42,6 +42,7 @@ use Friendica\Model\Notify\Type;
 use Friendica\Model\PermissionSet;
 use Friendica\Model\Profile;
 use Friendica\Model\User;
+use Friendica\Network\HTTPRequest;
 use Friendica\Network\Probe;
 use Friendica\Util\Crypto;
 use Friendica\Util\DateTimeFormat;
@@ -1221,7 +1222,7 @@ class DFRN
 
 		Logger::log('dfrn_deliver: ' . $url);
 
-		$curlResult = Network::curl($url);
+		$curlResult = HTTPRequest::curl($url);
 
 		if ($curlResult->isTimeout()) {
 			return -2; // timed out
@@ -1370,7 +1371,7 @@ class DFRN
 
 		Logger::log('dfrn_deliver: ' . "SENDING: " . print_r($postvars, true), Logger::DATA);
 
-		$postResult = Network::post($contact['notify'], $postvars);
+		$postResult = HTTPRequest::post($contact['notify'], $postvars);
 
 		$xml = $postResult->getBody();
 
@@ -1467,7 +1468,7 @@ class DFRN
 
 		$content_type = ($public_batch ? "application/magic-envelope+xml" : "application/json");
 
-		$postResult = Network::post($dest_url, $envelope, ["Content-Type: ".$content_type]);
+		$postResult = HTTPRequest::post($dest_url, $envelope, ["Content-Type: " . $content_type]);
 		$xml = $postResult->getBody();
 
 		$curl_stat = $postResult->getReturnCode();
