+++ src/Protocol/Salmon.php
@@ -22,6 +22,7 @@
 namespace Friendica\Protocol;
 
 use Friendica\Core\Logger;
+use Friendica\Network\HTTPRequest;
 use Friendica\Network\Probe;
 use Friendica\Util\Crypto;
 use Friendica\Util\Network;
@@ -72,7 +73,7 @@ class Salmon
 						$ret[$x] = substr($ret[$x], 5);
 					}
 				} elseif (Strings::normaliseLink($ret[$x]) == 'http://') {
-					$ret[$x] = Network::fetchUrl($ret[$x]);
+					$ret[$x] = HTTPRequest::fetchUrl($ret[$x]);
 				}
 			}
 		}
@@ -155,7 +156,7 @@ class Salmon
 		$salmon = XML::fromArray($xmldata, $xml, false, $namespaces);
 
 		// slap them
-		$postResult = Network::post($url, $salmon, [
+		$postResult = HTTPRequest::post($url, $salmon, [
 			'Content-type: application/magic-envelope+xml',
 			'Content-length: ' . strlen($salmon)
 		]);
@@ -180,7 +181,7 @@ class Salmon
 			$salmon = XML::fromArray($xmldata, $xml, false, $namespaces);
 
 			// slap them
-			$postResult = Network::post($url, $salmon, [
+			$postResult = HTTPRequest::post($url, $salmon, [
 				'Content-type: application/magic-envelope+xml',
 				'Content-length: ' . strlen($salmon)
 			]);
@@ -203,7 +204,7 @@ class Salmon
 			$salmon = XML::fromArray($xmldata, $xml, false, $namespaces);
 
 			// slap them
-			$postResult = Network::post($url, $salmon, [
+			$postResult = HTTPRequest::post($url, $salmon, [
 				'Content-type: application/magic-envelope+xml',
 				'Content-length: ' . strlen($salmon)]);
 			$return_code = $postResult->getReturnCode();
