+++ src/Util/Images.php
@@ -24,6 +24,7 @@ namespace Friendica\Util;
 use Friendica\Core\Logger;
 use Friendica\Core\System;
 use Friendica\DI;
+use Friendica\Network\HTTPRequest;
 use Imagick;
 
 /**
@@ -167,7 +168,7 @@ class Images
 			return $data;
 		}
 
-		$img_str = Network::fetchUrl($url, true, 4);
+		$img_str = HTTPRequest::fetchUrl($url, true, 4);
 
 		if (!$img_str) {
 			return [];
