+++ src/Worker/CheckVersion.php
@@ -24,6 +24,7 @@ namespace Friendica\Worker;
 use Friendica\Core\Logger;
 use Friendica\Database\DBA;
 use Friendica\DI;
+use Friendica\Network\HTTPRequest;
 use Friendica\Util\Network;
 
 /**
@@ -54,7 +55,7 @@ class CheckVersion
 		Logger::log("Checking VERSION from: ".$checked_url, Logger::DEBUG);
 
 		// fetch the VERSION file
-		$gitversion = DBA::escape(trim(Network::fetchUrl($checked_url)));
+		$gitversion = DBA::escape(trim(HTTPRequest::fetchUrl($checked_url)));
 		Logger::log("Upstream VERSION is: ".$gitversion, Logger::DEBUG);
 
 		DI::config()->set('system', 'git_friendica_version', $gitversion);
