+++ src/Worker/Directory.php
@@ -26,6 +26,7 @@ use Friendica\Core\Logger;
 use Friendica\Core\Worker;
 use Friendica\Database\DBA;
 use Friendica\DI;
+use Friendica\Network\HTTPRequest;
 use Friendica\Util\Network;
 
 /**
@@ -54,7 +55,7 @@ class Directory
 
 		Logger::log('Updating directory: ' . $arr['url'], Logger::DEBUG);
 		if (strlen($arr['url'])) {
-			Network::fetchUrl($dir . '?url=' . bin2hex($arr['url']));
+			HTTPRequest::fetchUrl($dir . '?url=' . bin2hex($arr['url']));
 		}
 
 		return;
